Route-benamingen aangepast naar consistentere en begrijpelijkere naamgeving
