+++ src/routes/products/[id]/+page.svelte
